style(coupon): 패키지명 오타 수정 (#250)
- 패키지명 excpeption을 exception으로 수정
