+++ src/main/java/com/firstone/greenjangteo/coupon/exception/message/NotFoundExceptionMessage.java
@@ -1,4 +1,4 @@
-package com.firstone.greenjangteo.coupon.excpeption.message;
+package com.firstone.greenjangteo.coupon.exception.message;
 
 public class NotFoundExceptionMessage {
     public static final String COUPON_NAME_NOT_FOUND_EXCEPTION = "해당 쿠폰 그룹이 존재하지 않습니다. 전송된 쿠폰 이름: ";
