+++ src/main/java/com/firstone/greenjangteo/coupon/exception/serious/AlreadyProvidedCouponException.java
@@ -1,10 +1,10 @@
-package com.firstone.greenjangteo.coupon.excpeption.serious;
+package com.firstone.greenjangteo.coupon.exception.serious;
 
 import com.firstone.greenjangteo.exception.AbstractSeriousException;
 import org.springframework.http.HttpStatus;
 
-public class InconsistentCouponSizeException extends AbstractSeriousException {
-    public InconsistentCouponSizeException(String message) {
+public class AlreadyProvidedCouponException extends AbstractSeriousException {
+    public AlreadyProvidedCouponException(String message) {
         super(message);
     }
