+++ src/main/java/com/firstone/greenjangteo/coupon/exception/serious/AlreadyUsedCouponException.java
@@ -1,4 +1,4 @@
-package com.firstone.greenjangteo.coupon.excpeption.serious;
+package com.firstone.greenjangteo.coupon.exception.serious;
 
 import com.firstone.greenjangteo.exception.AbstractSeriousException;
 import org.springframework.http.HttpStatus;
