+++ src/main/java/com/firstone/greenjangteo/coupon/exception/serious/InconsistentCouponSizeException.java
@@ -1,10 +1,10 @@
-package com.firstone.greenjangteo.coupon.excpeption.serious;
+package com.firstone.greenjangteo.coupon.exception.serious;
 
 import com.firstone.greenjangteo.exception.AbstractSeriousException;
 import org.springframework.http.HttpStatus;
 
-public class InsufficientRemainingQuantityException extends AbstractSeriousException {
-    public InsufficientRemainingQuantityException(String message) {
+public class InconsistentCouponSizeException extends AbstractSeriousException {
+    public InconsistentCouponSizeException(String message) {
         super(message);
     }
