+++ src/main/java/com/firstone/greenjangteo/coupon/model/Amount.java
@@ -4,8 +4,8 @@ import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 import java.util.Objects;
 
-import static com.firstone.greenjangteo.coupon.excpeption.message.BlankExceptionMessage.COUPON_AMOUNT_NO_VALUE_EXCEPTION;
-import static com.firstone.greenjangteo.coupon.excpeption.message.InvalidExceptionMessage.INVALID_COUPON_AMOUNT_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.BlankExceptionMessage.COUPON_AMOUNT_NO_VALUE_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.InvalidExceptionMessage.INVALID_COUPON_AMOUNT_EXCEPTION;
 import static com.firstone.greenjangteo.utility.RegularExpressionConstant.POSITIVE_INTEGER_PATTERN;
 
 public class Amount {
