+++ src/test/java/com/firstone/greenjangteo/coupon/model/ExpirationPeriodTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static com.firstone.greenjangteo.coupon.excpeption.message.BlankExceptionMessage.EXPIRATION_PERIOD_NO_VALUE_EXCEPTION;
-import static com.firstone.greenjangteo.coupon.excpeption.message.InvalidExceptionMessage.INVALID_EXPIRATION_PERIOD_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.BlankExceptionMessage.EXPIRATION_PERIOD_NO_VALUE_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.InvalidExceptionMessage.INVALID_EXPIRATION_PERIOD_EXCEPTION;
 import static com.firstone.greenjangteo.coupon.testutil.CouponTestConstant.EXPIRATION_PERIOD1;
 import static com.firstone.greenjangteo.coupon.testutil.CouponTestConstant.EXPIRATION_PERIOD2;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
