+++ src/test/java/com/firstone/greenjangteo/coupon/model/IssueQuantityTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static com.firstone.greenjangteo.coupon.excpeption.message.BlankExceptionMessage.ISSUE_QUANTITY_NO_VALUE_EXCEPTION;
-import static com.firstone.greenjangteo.coupon.excpeption.message.InvalidExceptionMessage.INVALID_ISSUE_QUANTITY_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.BlankExceptionMessage.ISSUE_QUANTITY_NO_VALUE_EXCEPTION;
+import static com.firstone.greenjangteo.coupon.exception.message.InvalidExceptionMessage.INVALID_ISSUE_QUANTITY_EXCEPTION;
 import static com.firstone.greenjangteo.coupon.testutil.CouponTestConstant.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
