+++ src/test/java/com/firstone/greenjangteo/coupon/model/entity/CouponGroupTest.java
@@ -1,8 +1,8 @@
 package com.firstone.greenjangteo.coupon.model.entity;
 
 import com.firstone.greenjangteo.application.model.CouponGroupModel;
-import com.firstone.greenjangteo.coupon.excpeption.serious.InconsistentCouponSizeException;
-import com.firstone.greenjangteo.coupon.excpeption.serious.InsufficientRemainingQuantityException;
+import com.firstone.greenjangteo.coupon.exception.serious.InconsistentCouponSizeException;
+import com.firstone.greenjangteo.coupon.exception.serious.InsufficientRemainingQuantityException;
 import com.firstone.greenjangteo.coupon.model.Amount;
 import com.firstone.greenjangteo.coupon.model.ExpirationPeriod;
 import com.firstone.greenjangteo.coupon.model.IssueQuantity;
@@ -25,7 +25,7 @@ import javax.persistence.EntityManager;
 import java.time.LocalDate;
 import java.util.List;
 
-import static com.firstone.greenjangteo.coupon.excpeption.message.AbnormalStateExceptionMessage.*;
+import static com.firstone.greenjangteo.coupon.exception.message.AbnormalStateExceptionMessage.*;
 import static com.firstone.greenjangteo.coupon.testutil.CouponTestConstant.*;
 import static com.firstone.greenjangteo.user.model.Role.ROLE_BUYER;
 import static com.firstone.greenjangteo.user.testutil.UserTestConstant.*;
